[#2] Rename: LoginButton 컴포넌트를 components 폴더로 이동
